backend/requirements.txt
# ===== requirements.txt =====
# Updated: October 2025 - Latest stable versions
fastapi==0.121.2
starlette==0.49.1
uvicorn[standard]==0.38.0
sqlalchemy==2.0.44
psycopg[binary]==3.2.12
pydantic==2.12.3
email-validator==2.3.0
python-multipart==0.0.20
openpyxl==3.1.5
reportlab==4.4.4
pydantic-settings==2.11.0
alembic==1.17.0
psutil==7.1.1
httpx==0.27.2
slowapi==0.1.9  # Rate limiting for API endpoints
fastapi-csrf-protect==1.0.7  # CSRF protection for state-changing operations
passlib[bcrypt]==1.7.4  # Password hashing
PyJWT==2.9.0            # JWT tokens (replaces python-jose)
docker==7.1.0
prometheus-client==0.21.0  # Prometheus metrics
prometheus-fastapi-instrumentator==7.0.0  # FastAPI Prometheus integration
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
redis==5.0.1  # Redis caching support (optional, enable via REDIS_ENABLED=true)
psycopg2-binary==2.9.9  # Fallback driver for postgresql:// connection strings
